Category
Criação da tabela categoria e produtos

+++ KarapinhaApi/src/main/java/tinario9954/gmail/com/KarapinhaApi/DTOS/RoleDTOS.java
@@ -12,13 +12,13 @@ import tinario9954.gmail.com.KarapinhaApi.Models.Role;
 public class RoleDTOS implements Serializable {
 
     private static final long serialVersionUID = 1L;
-    private Integer id;
+    private Long id;
     private String authority;
 
     public RoleDTOS() {
     }
 
-    public RoleDTOS(Integer id, String authority) {
+    public RoleDTOS(Long id, String authority) {
         super();
         this.id = id;
         this.authority = authority;
